using floats instead of doubles because.. well.. jack uses floats
git-svn-id: svn+ssh://yew/home/alex/code-repo/projects/datajockey/code/qtgui@467 be07c712-5235-0410-95d4-c99efef06bb2

+++ main.cpp
@@ -54,14 +54,14 @@ int main(int argc, char *argv[])
 	MixerChannelView * mixerChan = mixerPannel->mixerChannels()->front()->mixerChannel();
 	QObject::connect(
 			mixerChan,
-			SIGNAL(volumeChanged(double)),
-			mixerModel, SLOT(setVolume(double)));
+			SIGNAL(volumeChanged(float)),
+			mixerModel, SLOT(setVolume(float)));
 
 	QObject::connect(
 			mixerModel,
-			SIGNAL(volumeChanged(double)),
+			SIGNAL(volumeChanged(float)),
 			mixerChan,
-			SLOT(setVolume(double)));
+			SLOT(setVolume(float)));
 
 	QObject::connect(
 			mixerChan,
