+++ mixerchannelview.hpp
@@ -14,13 +14,13 @@ class MixerChannelView : public QWidget {
 		QSlider * volumeSlider();
 		QToolButton * muteButton();
 		void reset();
-		double volume() const;
+		float volume() const;
 		bool muted() const;
 	signals:
-		void volumeChanged(double volume);
+		void volumeChanged(float volume);
 		void mutedChanged(bool muted);
 	public slots:
-		void setVolume(double volume);
+		void setVolume(float volume);
 		void setVolume(int volume);
 		void setMuted(bool muted = true);
 	private:
